+++ .atomist/handlers/command/RestartBuild.ts
@@ -22,16 +22,10 @@ import {
     MappedParameters, Response, ResponseMessage,
 } from "@atomist/rug/operations/Handlers";
 
-import { handleErrors, wrap } from "@atomist/rugs/operations/CommonHandlers";
+import { wrap } from "@atomist/rugs/operations/CommonHandlers";
 import { execute } from "@atomist/rugs/operations/PlanUtils";
 
 import { travisSecretPath } from "./SecretPath";
-
-const buildIdParameter = {
-    displayName: "Build ID",
-    description: "Travis CI identifier of build to restart",
-    pattern: "^\\d+$",
-};
 
 /**
  * Command handler for restarting a Travis CI build that responds to
@@ -49,7 +43,11 @@ const buildIdParameter = {
 @Intent("restart build", "restart travis build")
 class RestartBuild implements HandleCommand {
 
-    @Parameter(buildIdParameter)
+    @Parameter({
+        displayName: "Build ID",
+        description: "Travis CI identifier of build to restart",
+        pattern: "^\\d+$",
+    })
     public buildId: string;
 
     @MappedParameter(MappedParameters.GITHUB_REPOSITORY)
@@ -61,63 +59,11 @@ class RestartBuild implements HandleCommand {
     public handle(ctx: HandlerContext): CommandPlan {
         const plan = new CommandPlan();
 
-        const GITHUB_API_BASE = "https://api.github.com";
-        const headers = {
-            "Accept": "application/vnd.github.v3+json",
-            "Authorization": `token #{${travisSecretPath}}`,
-            "Content-Type": "application/json",
-        };
-        const response: CommandRespondable<Execute> = {
-            instruction: {
-                kind: "execute",
-                name: "http",
-                parameters: {
-                    method: "get",
-                    url: `${GITHUB_API_BASE}/repos/${this.owner}/${this.repo}`,
-                    config: { headers },
-                },
-            },
-            onSuccess: {
-                kind: "respond",
-                name: "SendRestart",
-                parameters: { buildId: this.buildId },
-            },
-        };
-        plan.add(handleErrors(response));
+        plan.add(wrap(execute("travis-restart-build", this),
+            `Successfully restarted ${this.owner}/${this.repo} Travis CI build ${this.buildId}`, this));
 
         return plan;
     }
 }
 
 export const restartBuild = new RestartBuild();
-
-/**
- * Response handler that parses the response to the GitHub repo API to
- * extract the visibility of the repo and then call the
- * `restart-travis-build` Rug function.
- *
- * @param buildId  integer ID of Travis CI build to restart
- */
-@ResponseHandler("SendRestart", "send restart to Travis CI API")
-@Tags("travis-ci", "ci")
-@Secrets(travisSecretPath)
-export class SendRestart implements HandleResponse<any> {
-
-    @Parameter(buildIdParameter)
-    public buildId: string;
-
-    public handle( @ParseJson response: Response<any>): CommandPlan {
-        const plan = new CommandPlan();
-
-        const repo = response.body;
-        const privateKey = "private";
-        const isPrivate = repo[privateKey] as boolean;
-        const visibility = (isPrivate) ? "private" : "public";
-        plan.add(wrap(execute("restart-travis-build", { buildId: this.buildId, visibility }),
-            `Successfully restarted Travis CI build ${this.buildId}`, this));
-
-        return plan;
-    }
-}
-
-export const sendRestart = new SendRestart();